requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
openai-whisper==20231117
librosa==0.10.1
pydub==0.25.1
SpeechRecognition==3.10.0
numpy==1.24.3
